+++ src/autocomplete/overlay/autocomplete-ref.ts
@@ -1,7 +1,4 @@
-import { ThyAbstractInternalOverlayRef, ThyAbstractOverlayRef } from 'ngx-tethys/core';
-
-import { OverlayRef } from '@angular/cdk/overlay';
-
+import { ThyAbstractInternalOverlayRef, ThyAbstractOverlayOptions, ThyAbstractOverlayRef } from 'ngx-tethys/core';
 import { ThyAutocompleteContainer } from './autocomplete-container.component';
 import { ThyAutocompleteConfig } from './autocomplete.config';
 import { autocompleteAbstractOverlayOptions } from './autocomplete.options';
@@ -9,11 +6,15 @@ import { autocompleteAbstractOverlayOptions } from './autocomplete.options';
 export abstract class ThyAutocompleteRef<T, TResult = any> extends ThyAbstractOverlayRef<T, ThyAutocompleteContainer, TResult> {}
 
 export class ThyInternalAutocompleteRef<T, TResult = any>
-    extends ThyAbstractInternalOverlayRef<T, ThyAutocompleteContainer, TResult>
+    extends ThyAbstractInternalOverlayRef<T, ThyAutocompleteContainer, TResult, ThyAutocompleteConfig>
     implements ThyAutocompleteRef<T, TResult>
 {
-    constructor(overlayRef: OverlayRef, containerInstance: ThyAutocompleteContainer, config: ThyAutocompleteConfig) {
-        super(autocompleteAbstractOverlayOptions, overlayRef, containerInstance, config);
+    constructor() {
+        super();
+    }
+
+    protected get options(): ThyAbstractOverlayOptions {
+        return autocompleteAbstractOverlayOptions;
     }
 
     /**